Move connection handler tests to the right module
