+++ tests/test_connection_handler.py
@@ -1,3 +1,4 @@
+import json
 import pathlib
 import unittest
 
@@ -89,6 +90,32 @@ class TestConnectionHandler(unittest.TestCase):
             }
         )
         self.assertIsInstance(connection, Connection)
+
+    def test_connection_handler_no_ingress_port(self):
+        with open(self.CONNECTION_FILE_P2P, "r", encoding="utf-8") as f:
+            connection_data = json.load(f)
+
+        connection_data["ingress_port"] = None
+
+        self.assertRaisesRegex(
+            MissingAttributeException,
+            f"Missing attribute 'ingress_port' while parsing <{connection_data}>",
+            ConnectionHandler().import_connection_data,
+            connection_data,
+        )
+
+    def test_connection_handler_no_egress_port(self):
+        with open(self.CONNECTION_FILE_P2P, "r", encoding="utf-8") as f:
+            connection_data = json.load(f)
+
+        connection_data["egress_port"] = None
+
+        self.assertRaisesRegex(
+            MissingAttributeException,
+            f"Missing attribute 'egress_port' while parsing <{connection_data}>",
+            ConnectionHandler().import_connection_data,
+            connection_data,
+        )
 
 
 if __name__ == "__main__":
